Refactoring specs
