+++ spec/features/attachment/add_file_for_answer_spec.rb
@@ -7,8 +7,8 @@ feature 'add files to answer', :js do
 
   context 'creates answer' do
     it_behaves_like 'files features' do
-      given(:filable_selector) { '.new-answer' }
-      given(:filable_btn) { click_on 'Post Your Answer' }
+      given(:file_selector) { '.new-answer' }
+      given(:file_btn) { click_on 'Post Your Answer' }
       given(:background_info) do
         visit question_path(question)
         fill_in 'Your Answer:', with: 'Test text'
@@ -18,8 +18,8 @@ feature 'add files to answer', :js do
 
   context 'edits answer' do
     it_behaves_like 'files features' do
-      given(:filable_selector) { "#answer-#{answer.id}" }
-      given(:filable_btn) { click_on 'Update Your Answer' }
+      given(:file_selector) { "#answer-#{answer.id}" }
+      given(:file_btn) { click_on 'Update Your Answer' }
       given(:background_info) do
         visit question_path(question)
         click_on 'Edit Answer'
